% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetuner.R
\name{check_env_initialized}
\alias{check_env_initialized}
\title{Helper function to check if the Python environment is initialized.}
\usage{
check_env_initialized()
}
\description{
Helper function to check if the Python environment is initialized.
}
