% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetuner.R
\name{finetune_model}
\alias{finetune_model}
\title{Fine-Tune a Sequence Model}
\usage{
finetune_model(
  datasets,
  task_type = "classification",
  model_name,
  training_args,
  num_labels = NULL,
  use_cuda = FALSE
)
}
\arguments{
\item{datasets}{A list of datasets created by `prepare_finetuning_data`.}

\item{task_type}{A string, either `"classification"` or `"regression"`.}

\item{model_name}{The name of the pre-trained model from Hugging Face Hub.}

\item{training_args}{A `TrainingArguments` object from `transformers`.}

\item{num_labels}{For classification, the number of unique classes. For
regression, this is automatically set to 1 and can be omitted.}

\item{use_cuda}{If `TRUE`, will attempt to use a GPU if available.}
}
\value{
A list with the trained `Trainer` and prediction output.
}
\description{
Orchestrates the fine-tuning process for classification or regression.
}
