% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetuner.R
\name{get_python_dataset_class}
\alias{get_python_dataset_class}
\title{Get Python Dataset Class}
\usage{
get_python_dataset_class()
}
\value{
A Python class object (`RDataset`).
}
\description{
This internal function defines a Python torch.utils.data.Dataset class
using a string.
}
